telegram_bot/requirements.txt
python-telegram-bot==20.3
requests==2.31.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==5.0.0